Sync with AppHeader
Slides in and out of viewport in sync with AppHeader (via headroom.js).
Uses a MutationObserver to track when AppHeader is toggled (watches css
classes) and sets visible state accordingly.

Uses some scroll hackery at the top of the document to smooth the
transition, and eventually toggle 'position' from 'relative' to 'fixed'.
It's grossly inefficient as the onscroll callback is not debounced or
throttled, but cpu/ram spikes were minimal (on my laptop) so whatever.

I tried some naive debounce/throttle code found on the interwebs, but
they would spin the fans up if I scrolled alot. I guess sometimes simple
is the best.

The RIGHT way to do all of this would be to move some state and
functions to the store and then merge the MemorialTabs (a secondary
AppHeader, more or less) with the actual AppHeader, so that it's all
under the control of the 'headroom.js' component.

But what's the fun in that? If I had just refactored the code, I would
never have gone down the rabit hole to learn about
debouncing/throttling, requestAnimationFrame(), MutationObservers, and a
whole bunch of other stuff about DOM events. And the hacky version works
pretty well. I'd go as far to say I'm rather proud of it.