Create, edit, delete functionality for Chronicle
Right now all of this is kind of a mess. But it works, so I'm going
forward. Things are quite tangled up.

Edit modal is attached to ChronicleCard, forks it's own state from the
props that ChronicleCard passes in. Component is destroyed each time,
thanks to having a unique key.

Add modal is attached to the ChronicleFab. The logic behind this was to
attach the components to the ones that call/spawn them. Clicking the Fab
brings up the AddItemModal, clicking through the ChronicleCard, to edit
item brings up the EditItemModal, etc.

Would probably be more sane to treat all of these things as different
"screens", have them all as top level components within Memorial, and
move as much state to the store as possible. That's on the todo list.

Gotta say I really don't like passing props between components. I think
I would be less crabby if almost everything was in the store and
components only grabbed what they needed. But I'm not *that* crabby now,
so we'll see if I ever get around to it.